+++ traveldesign/package.json
@@ -12,6 +12,7 @@
     "@strapi/blocks-react-renderer": "^1.0.2",
     "bootstrap": "^5.3.5",
     "next": "15.3.1",
+    "next-intl": "^4.1.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-icons": "^5.5.0",
