--- traveldesign/src/app/(pages)/experiencias/uploader/page.jsx
@@ -1,16 +1,0 @@
-"use client";
-
-import XlsUploader from "@/app/components/XlsUploader";
-
-const Index = () => {
-  return (
-    <div>
-      <h1>Uploader XLS</h1>
-      <div>
-        <XlsUploader />
-      </div>
-    </div>
-  );
-};
-
-export default Index;
