+++ traveldesign/src/app/[locale]/(pages)/experiencias/uploader/page.jsx
@@ -1,0 +1,16 @@
+"use client";
+
+import XlsUploader from "@/app/[locale]/components/XlsUploader";
+
+const Index = () => {
+  return (
+    <div>
+      <h1>Uploader XLS</h1>
+      <div>
+        <XlsUploader />
+      </div>
+    </div>
+  );
+};
+
+export default Index;
