+++ traveldesign/src/app/[locale]/components/PartnersSection.jsx
