+++ traveldesign/src/app/[locale]/globals.css
