+++ traveldesign/src/app/[locale]/layout.jsx
@@ -1,0 +1,41 @@
+import { NextIntlClientProvider, hasLocale } from "next-intl";
+import { notFound } from "next/navigation";
+import { routing } from "@/i18n/routing";
+
+import "bootstrap/dist/css/bootstrap.min.css";
+import { Poppins } from "next/font/google";
+import Header from "@/app/[locale]/components/Header";
+import Footer from "@/app/[locale]/components/Footer";
+import BootstrapProvider from "@/app/[locale]/providers/bootstrap-provider";
+import "./globals.css";
+
+const poppins = Poppins({
+  subsets: ["latin"],
+  weight: ["400", "600", "700"],
+  display: "swap",
+});
+
+export const metadata = {
+  title: "Travel Design",
+  description: "Luxury travel experiences",
+};
+
+export default async function RootLayout({ children, params }) {
+  const { locale } = await params;
+  if (!hasLocale(routing.locales, locale)) {
+    notFound();
+  }
+
+  return (
+    <html lang={locale}>
+      <body className={`${poppins.variable} antialiased`}>
+        <BootstrapProvider />
+        <NextIntlClientProvider>
+          <Header />
+          <main>{children}</main>
+          <Footer />
+        </NextIntlClientProvider>
+      </body>
+    </html>
+  );
+}
