--- traveldesign/src/app/lib/experiences/get-menu-experiences.js
@@ -1,9 +1,0 @@
-import { query } from "../strapi";
-
-export function getMenuExperiences(profile) {
-  return query(
-    `experience-categories?filters[profile][$eq]=${profile}&locale=es-MX&populate=experiences`
-  ).then((res) => {
-    return res.data;
-  });
-}
